+++ src/refactoring/features/coupon/ui/index.ts
@@ -1,1 +1,2 @@
 export { default as CouponSelect } from "./CouponSelect";
+export { default as CreateCouponForm } from "./CreateCouponForm";
